Sprint 4 done.

+++ Models/ProductInventory.cs
@@ -4,7 +4,7 @@ namespace GlamourHub.Models
 {
     public class ProductInventory
     {
-        [Key] public int ProductId { get; set; }
+        [Key] public Int64 ProductId { get; set; }
         public string ProductName { get; set; }
         public int InStock { get; set; }
         public int SoldQuantity { get; set; }
